third_party/blink/web_tests/external/wpt/css/css-values/random-in-container-query.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] random() should be disallowed in @container style() query
  assert_equals: expected "true" but got "false"
[FAIL] random() in var() should be disallowed in @container style() query
  assert_equals: expected "true" but got "false"
Harness: the test ran to completion.
